requirements.txt
fastapi==0.124.4
fastapi[standard]

uvicorn==0.38.0

asyncpg==0.31.0

redis==7.1.0
rq==2.6.1

pydantic==2.11.9

python-dotenv==1.1.1